Refactors LocalValueStore to delegate map-related state management to MapStateRepository
