+++ app/src/main/java/org/groundplatform/android/repository/MapStateRepository.kt
@@ -17,7 +17,10 @@ package org.groundplatform.android.repository
 
 import javax.inject.Inject
 import javax.inject.Singleton
-import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
 import org.groundplatform.android.persistence.local.LocalValueStore
 import org.groundplatform.android.ui.map.CameraPosition
 import org.groundplatform.android.ui.map.MapType
@@ -26,11 +29,25 @@ import org.groundplatform.android.ui.map.MapType
 @Singleton
 class MapStateRepository @Inject constructor(private val localValueStore: LocalValueStore) {
 
-  val mapTypeFlow: Flow<MapType> by localValueStore::mapTypeFlow
-  var mapType: MapType by localValueStore::mapType
+  private val _mapType = MutableStateFlow(mapType)
+  val mapTypeFlow: StateFlow<MapType> = _mapType.asStateFlow()
+  var mapType: MapType
+    get() = MapType.entries[localValueStore.mapType]
+    set(value) {
+      localValueStore.mapType = value.ordinal
+      _mapType.update { value }
+    }
+
+  private val _offlineImageryEnabled = MutableStateFlow(isOfflineImageryEnabled)
+  val offlineImageryEnabledFlow: StateFlow<Boolean> = _offlineImageryEnabled.asStateFlow()
+  var isOfflineImageryEnabled: Boolean
+    get() = localValueStore.isOfflineImageryEnabled
+    set(value) {
+      localValueStore.isOfflineImageryEnabled = value
+      _offlineImageryEnabled.update { value }
+    }
+
   var isLocationLockEnabled: Boolean by localValueStore::isLocationLockEnabled
-  val offlineImageryEnabledFlow: Flow<Boolean> by localValueStore::offlineImageryEnabledFlow
-  var isOfflineImageryEnabled: Boolean by localValueStore::isOfflineImageryEnabled
 
   fun setCameraPosition(cameraPosition: CameraPosition) =
     localValueStore.setLastCameraPosition(localValueStore.lastActiveSurveyId, cameraPosition)
